refactor: update & improve usage of `constants.ts` files
Expanding the usage of package-local `constants.ts` files to not just be
a barrel file for project-wide constants but also for defining and
sharing local constants.
