+++ src/clients/constants.ts
@@ -1,4 +1,4 @@
-import { INPUT_NAME_REPO_TOKEN } from "../constants";
+const INPUT_NAME_REPO_TOKEN = "repo-token";
 
 export {
   INPUT_NAME_REPO_TOKEN,
