+++ src/constants/index.ts
@@ -1,2 +1,1 @@
 export * from "./events";
-export * from "./inputs";
