--- src/constants/inputs.ts
@@ -1,28 +1,0 @@
-const DEFAULT_IGNORE_GLOBS = [];
-const DEFAULT_IS_DRY_RUN = false;
-const DEFAULT_IS_STRICT_MODE = false;
-const DEFAULT_SVGO_V1_CONFIG_PATH = ".svgo.yml";
-const DEFAULT_SVGO_V2_CONFIG_PATH = "svgo.config.js";
-const DEFAULT_SVGO_VERSION = 2;
-
-const INPUT_NAME_DRY_RUN = "dry-run";
-const INPUT_NAME_IGNORE = "ignore";
-const INPUT_NAME_REPO_TOKEN = "repo-token";
-const INPUT_NAME_STRICT = "strict";
-const INPUT_NAME_SVGO_CONFIG = "svgo-config";
-const INPUT_NAME_SVGO_VERSION = "svgo-version";
-
-export {
-  DEFAULT_IGNORE_GLOBS,
-  DEFAULT_IS_DRY_RUN,
-  DEFAULT_IS_STRICT_MODE,
-  DEFAULT_SVGO_V1_CONFIG_PATH,
-  DEFAULT_SVGO_V2_CONFIG_PATH,
-  DEFAULT_SVGO_VERSION,
-  INPUT_NAME_DRY_RUN,
-  INPUT_NAME_IGNORE,
-  INPUT_NAME_REPO_TOKEN,
-  INPUT_NAME_STRICT,
-  INPUT_NAME_SVGO_CONFIG,
-  INPUT_NAME_SVGO_VERSION,
-};
