+++ src/file-systems/constants.ts
@@ -1,0 +1,8 @@
+const LIST_FILES_ALWAYS_IGNORE: string[] = [
+  ".git",
+  "node_modules",
+];
+
+export {
+  LIST_FILES_ALWAYS_IGNORE,
+};
