+++ src/filters/constants.ts
@@ -1,0 +1,5 @@
+const STATUS_REMOVED = "removed";
+
+export {
+  STATUS_REMOVED,
+};
